Normalize small javaagent package abbreviations
