+++ instrumentation/kotlinx-coroutines/kotlinx-coroutines-flow-1.3/javaagent-kotlin/src/main/kotlin/io/opentelemetry/javaagent/instrumentation/kotlinx/coroutines/flow/v1_3/FlowUtil.kt
@@ -1,0 +1,20 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.javaagent.instrumentation.kotlinx.coroutines.flow.v1_3
+
+import io.opentelemetry.context.Context
+import io.opentelemetry.instrumentation.api.instrumenter.Instrumenter
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.onCompletion
+
+fun <REQUEST, RESPONSE, T> onComplete(
+  flow: Flow<T>,
+  instrumenter: Instrumenter<REQUEST, RESPONSE>,
+  context: Context,
+  request: REQUEST & Any,
+): Flow<T> = flow.onCompletion { cause: Throwable? ->
+  instrumenter.end(context, request, null, cause)
+}
