+++ instrumentation/scala-fork-join-2.8/javaagent/src/test/scala/io/opentelemetry/javaagent/instrumentation/scala/fork/join/v2_8/ForkJoinPoolBridge.scala
@@ -1,0 +1,45 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.javaagent.instrumentation.scala.fork.join.v2_8
+
+import java.util.concurrent._
+import scala.concurrent.forkjoin.{ForkJoinPool, ForkJoinTask}
+
+class ForkJoinPoolBridge(delegate: ForkJoinPool)
+    extends AbstractExecutorService {
+
+  override def shutdown(): Unit = delegate.shutdown()
+
+  override def shutdownNow(): java.util.List[Runnable] = delegate.shutdownNow()
+
+  override def isShutdown: Boolean = delegate.isShutdown
+
+  override def isTerminated: Boolean = delegate.isTerminated
+
+  override def awaitTermination(timeout: Long, unit: TimeUnit): Boolean =
+    delegate.awaitTermination(timeout, unit)
+
+  override def submit[T](task: Callable[T]): Future[T] = delegate.submit(task)
+
+  override def submit[T](task: Runnable, result: T): Future[T] =
+    delegate.submit(task, result)
+
+  override def submit(task: Runnable): Future[_] = delegate.submit(task)
+
+  override def invokeAll[T](
+      tasks: java.util.Collection[_ <: Callable[T]]
+  ): java.util.List[Future[T]] =
+    delegate.invokeAll(tasks)
+
+  override def invokeAny[T](tasks: java.util.Collection[_ <: Callable[T]]): T =
+    super.invokeAny(tasks)
+
+  def invoke[T](task: ForkJoinTask[T]): T = delegate.invoke(task)
+
+  override def execute(command: Runnable): Unit = delegate.execute(command)
+
+  def execute[T](task: ForkJoinTask[T]): Unit = delegate.execute(task)
+}
